Sprint 4 Parte I terminada

+++ ProyectoGerencia/ProyectoGerencia/ViewModels/Pago/DetallesVM.cs
@@ -11,6 +11,6 @@ namespace ProyectoGerencia.ViewModels.Pago
         public PersonaJuridica PersonaJuridica;
 
         public double SaldoTotal;
-        public double FacturaId;
+        public int FacturaId;
     }
 }
